+++ artifex/models/text_summarization/__init__.py
@@ -1,5 +1,6 @@
 from .text_summarization import TextSummarization
 
+
 __all__ = [
     "TextSummarization"
 ]